inference/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
flask>=2.0.0
pillow>=8.0.0
